requirements.txt
google-auth==2.23.3
google-api-python-client==2.102.0
google-cloud-vision==3.4.4
google-cloud-storage==2.10.0
google-auth-oauthlib==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3
pytesseract==0.3.10
opencv-python==4.8.1.78 